Add files via upload
This notebook reads in sorted_commits.csv and then produces a list of commits by day. The list has an entry for everyday from the starting date (0th row in table) to ending date (-1 row in table). For days of no commits, it fills in a zero. Finally, it writes out the list to  commit_counts.txt.