Update patterns.py
Some diffs have a + or - followed immediately with a part of the line. I changed the code to check for things like ++ or --. But +def, etc. is ok. Also, I am not so sure about the check if the line length is >1. I have seen single + or -, e.g., +\n or -\n. When you split into lines, you get a single + or -. Don't these represent newlines being added or deleted?