Update patterns.py
Trying to make process_single_commit smarter about parsing changes. As noted in a comment, there still is a kluge with single + or - not followed by any text. I believe these represent the adding or deleting of blank lines in the file, e.g., for whitespace readability. I am adding a dummy space following these types of changes to they are not lost.